USUM Support

+++ PKHeX.WinForms/Util/ConfigUtil.cs
@@ -5,7 +5,7 @@ namespace PKHeX.WinForms
 {
     public static class ConfigUtil
     {
-        public static bool checkConfig()
+        public static bool CheckConfig()
         {
             try
             {
@@ -15,7 +15,7 @@ namespace PKHeX.WinForms
             catch (ConfigurationErrorsException e)
             {
                 string path = (e.InnerException as ConfigurationErrorsException)?.Filename;
-                if (path != null)
+                if (path != null && File.Exists(path))
                     File.Delete(path);
                 return false;
             }
